fix: cover ambient declarations and split JSDoc checker modules
